style.css: Reformat and improve checkbox accessibility
Previously the expandable checkboxes were { display: none; } which made
them impossible to focus via the keyboard. Here we change it to {
opacity: 0; position: absolute; }, so it becomes keyboard-focusable.
Then we add an outline to the <label> when the checkbox is focused so
users could visually identify that the checkbox has been focused.